campapp: RGB LED amplifying
Brightness of RGB LED animations varies a lot, and sometimes
desired brightness depends on the environment and situation.
Aside of editing the animations directly, there was no way to
adjust the brightness of the LEDs.

This commit introduces a new config entry, "rgbleds_amp", which
allows to adjust the brightness of RGB LEDs in campapp animations.

The default value of 8 is a neutral setting.

Keep in mind that with static animations amp setting does not
apply immediately (either the frame delay must pass or the new
animation has to be loaded).